eox-report/requirements.txt
-i https://pypi.org/simple
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
intersight==1.0.9.6484
pem==21.2.0
pycryptodome==3.15.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
urllib3==1.26.9
xlsxwriter==3.0.3
